fix: (provider) update resource
Update resource was not working due to the extra ResourceRead at the beginning of each update.
I replaced every TypeSet with TypeList because during the apply process every element of the set is removed and added again which causes the field id to go null. TypeList only modify the already existing elements and does not remove them.

Signed-off-by: Alexandre Gomez <gomez.a.corneille@gmail.com>